# v2/api/swagger.yml
openapi: 3.0.0
info:
  description: |
    This Hermez node API is the layer that allows 3rd party apps and services to interface with the node to use the layer two features of the Hermez rollup.
    Example of these apps are:
    * Wallet: send L2 transactions, check balance, ...
    * Explorer: List transactions, slots, batches, ...
    * Exchange integrations

    Note that some of the interactions with the rollup must be done using the Ethereum network directly.
    Another way to integrate with the rollup is to [deploy a node](github.com/arnaubennassar/hermez-node/v2/) and connect directly to its PostgreSQL database.
    
    ### Pagination
    
    #### Usage

    All the endpoints that return a list of undefined size use pagination. Unless the opposite is explicitly said.
    In order to use pagination, three query parameters are used:
    * `fromItem`: indicates the first item to be returned. In general, this parameter shouldn't be provided in the first call to the endpoint, and use the `itemId` of the last returned item (+/-) 1, if the order is (ascending/descending).
    * `order`: all paginated items are ordered chronologically. However, the specific fields to guarantee this order depend on each endpoint. For this purpose, `itemId` is used (itemId follows ascending chronological order except for unforged L1 user transactions). If the parameter is not provided, ascending order will be used by default. 
...
    
...

    #### Reorgs and Safety

...

    ### Signatures

...

...

...

...

...
servers:
...
paths:
...
  '/accounts':
    get:
      tags:
        - Explorer
      summary: Get accounts balances and other associated information.
      description: Get account balances and other associated information.
      operationId: getAccounts
      parameters:
        - name: hezEthereumAddress
          in: query
          description: Only get accounts associated to an Ethereum address. Incompatible with the query `BJJ`.
          required: false
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: BJJ
          in: query
          description: Only get accounts associated to a BabyJubJub compressed public key.  Incompatible with the query `hezEthereumAddress`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: tokenIds
          in: query
          required: false
          description: Only get accounts of specific tokens. This is represented by a comma separated list of token identifiers.
          schema:
            type: string
            description: Comma separated list of token identifiers.
            example: "3,87,91"
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Accounts will be ordered by increasing account index.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounts'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/accounts/{accountIndex}':
    get:
      tags:
        - Explorer
      summary: Get an account by its index.
      description: Get an account by its index.
      operationId: getAccount
      parameters:
        - name: accountIndex
          in: path
          description: Identifier of an account.
          required: true
          schema:
            $ref: '#/components/schemas/AccountIndex'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
  '/transactions-pool':
    post:
      tags:
        - Coordinator
      summary: Add an L2 transaction to the coordinator's pool.
      description: >-
        Send L2 transaction. The transaction will be stored in the transaction pool of the coordinator and eventually forged.
      operationId: postTx
      requestBody:
        description: Signed transaction.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPoolL2Transaction'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionId'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
    get:
      tags:
        - Coordinator
      summary: Get transactions that are in the pool.
      operationId: getPoolTxs
      parameters:
        - name: state
          in: query
          required: false
          description: State of the transactions, e.g. "pend"
          schema:
            $ref: '#/components/schemas/PoolL2TransactionState'
        - name: tokenId
          in: query
          required: false
          description: Only get transactions of specific token
          schema:
            $ref: '#/components/schemas/TokenId'
        - name: accountIndex
          in: query
          required: false
          description: Only get transactions sent from and to a specific account. Incompatible with the queries `fromAccountIndex`, `toAccountIndex`, `hezEthereumAddress`, `fromHezEthereumAddress`, `toHezEthereumAddress`, `BJJ`, `toBJJ`, `fromBJJ`.
          schema:
            $ref: '#/components/schemas/AccountIndex'
        - name: fromAccountIndex
          in: query
          required: false
          description: Only get transactions sent from a specific account. Incompatible with the queries `accountIndex`, `hezEthereumAddress`, `fromHezEthereumAddress`, `toHezEthereumAddress`, `BJJ`, `toBJJ`, `fromBJJ`.
          schema:
            $ref: '#/components/schemas/AccountIndex'
        - name: toAccountIndex
          in: query
          required: false
          description: Only get transactions sent to a specific account. Incompatible with the queries `accountIndex`, `hezEthereumAddress`, `fromHezEthereumAddress`, `toHezEthereumAddress`, `BJJ`, `toBJJ`, `fromBJJ`.
          schema:
            $ref: '#/components/schemas/AccountIndex'
        - name: type
          in: query
          required: false
          description: Only get transactions of a specific type
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: hezEthereumAddress
          in: query
          required: false
          description: Only get transactions sent from or to an account associated with an Ethereum address. Incompatible with the queries `BJJ`, `fromBJJ`, `toBJJ` and `accountIndex`, `fromAccountIndex`, `toAccountIndex`, `toHezEthereumAddress`, `fromHezEthereumAddress`.
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: fromHezEthereumAddress
          in: query
          required: false
          description: Only get transactions sent from an account associated with an Ethereum address. Incompatible with the queries `BJJ`, `fromBJJ`, `toBJJ` and `accountIndex`, `fromAccountIndex`, `toAccountIndex`, `hezEthereumAddress`.
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: toHezEthereumAddress
          in: query
          required: false
          description: Only get transactions sent to an account associated with an Ethereum address. Incompatible with the queries `BJJ`, `fromBJJ`, `toBJJ` and `accountIndex`, `fromAccountIndex`, `toAccountIndex`, `hezEthereumAddress`.
          schema:
            $ref: '#/components/schemas/HezEthereumAddress'
        - name: BJJ
          in: query
          description: Only get transactions associated with a BabyJubJub compressed public key. The BJJ is in base 64. Incompatible with the queries `hezEthereumAddress` and `accountIndex`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: fromBJJ
          in: query
          description: Only get transactions sent from a BabyJubJub compressed public key. Incompatible with the queries `hezEthereumAddress`, `fromHezEthereumAddress`, `toHezEthereumAddress` and `accountIndex`, `fromAccountIndex`, `toAccountIndex`, `BJJ`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: toBJJ
          in: query
          description: Only get transactions sent to a BabyJubJub compressed public key. Incompatible with the queries `hezEthereumAddress`, `fromHezEthereumAddress`, `toHezEthereumAddress` and `accountIndex`, `fromAccountIndex`, `toAccountIndex`, `BJJ`.
          required: false
          schema:
            $ref: '#/components/schemas/BJJ'
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Transactions will be ordered by item_id.
          schema:
            type: string
            default: ASC
            enum:
              - ASC
              - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolL2Transactions'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/transactions-pool/{id}':
    get:
      tags:
        - Coordinator
      summary: Get details and status of a transaction that is in the pool.
      description: >-
        Get transaction from the pool by its ID. This endpoint is specially useful for tracking the status of a transaction that may not be forged yet.
        Only transactions from the pool will be returned.
        Note that the transaction pool is different for each coordinator and therefore only a coordinator that has received a specific transaction
        will be able to provide information about that transaction.
      operationId: getPoolTx
      parameters:
        - name: id
          in: path
          description: Transaction identifier.
          required: true
          schema:
            $ref: '#/components/schemas/TransactionId'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolL2Transaction'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
components:
  schemas:
    ItemId:
      type: integer
      description: Position of the item in the DB. This is useful for pagination, but has  nothing to do with the protocol.
    PostPoolL2Transaction:
      type: object
      description: L2 transaction to be posted.
      properties:
        id:
          $ref: '#/components/schemas/TransactionId'
        type:
          $ref: '#/components/schemas/TransactionTypeL2'
        tokenId:
          $ref: '#/components/schemas/TokenId'
        fromAccountIndex:
          $ref: '#/components/schemas/AccountIndex'
        toAccountIndex:
          type: string
          description: >-
            Identifier of the destination account. It references the position where the account is inside the state Merkle tree.
            The identifier is built using: `hez:` + `token symbol:` + `index`. If this is provided, toHezEthereumAddress and toBjj
            must be null. To perform an exit the value hez:EXIT:1 must be used.
          example: null
          nullable: true
        toHezEthereumAddress:
          type: string
          description: "Address of an Ethereum account linked to the Hermez Network. If this is provided, toAccountIndex and toBjj must be null."
          pattern: "^hez:0x[a-fA-F0-9]{40}$"
          example: "hez:0xaa942cfcd25ad4d90a62358b0dd84f33b398262a"
          nullable: true
        toBjj:
          type: string
          description: >-
            BabyJubJub compressed public key, encoded as base64 URL (RFC 4648), which result in 33 bytes. The padding byte is replaced by a sum of the encoded bytes.
            If this is provided, toAccountIndex must be null and toHezEthereumAddress must be hez:0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF.
          pattern: "^hez:[A-Za-z0-9_-]{44}$"
          example: null
          nullable: true
        amount:
          allOf:
            - $ref: '#/components/schemas/BigInt'
            - description: Amount of tokens to be sent.
              example: "6300000000000000000"
        fee:
          $ref: '#/components/schemas/FeeSelector'
        nonce:
          $ref: '#/components/schemas/Nonce'
        signature:
          allOf:
            - $ref: '#/components/schemas/BJJSignature'
            - description: Signature of the transaction. More info [here](https://idocs.hermez.io/#/spec/zkrollup/README?id=l2a-idl2).
            - example: "72024a43f546b0e1d9d5d7c4c30c259102a9726363adcc4ec7b6aea686bcb5116f485c5542d27c4092ae0ceaf38e3bb44417639bd2070a58ba1aa1aab9d92c03"
        maxNumBatch:
          type: integer
          description: Batch until the transaction can be forged. If the transaction isn't forged before batch `maxNumBatch`, it will never be forged. 0 means no restriction.
          example: 0  
        requestOffset:
          type: integer
          description: |
            Identifier of the relative index of the transaction that would be linked. 0 means no linked transaction.
            More info on how the identifiers are built [here](https://docs.hermez.io/#/developers/protocol/hermez-protocol/circuits/circuits?id=rq-tx-verifier)
          example: 0
          nullable: true
      example:
        id: '0x020000000001000000000006'
        type: Transfer
        tokenId: 6
        fromAccountIndex: hez:DAI:256
        toAccountIndex: hez:DAI:257
        toHezEthereumAddress: 
        toBjj: 
        amount: '100000000000000'
        fee: 0
        nonce: 6
        signature: 1a79dd5e661d58266901a0de8afb046b466c4c1af937100f627a421771f2911fa3fde8ea2e272b4802a8b1f1229689292acd6f7e8ab4cadc4ab37b6b9e13a101
        maxNumBatch: 0
        requestOffset: 0
      additionalProperties: false
      required:
      - id
      - type
      - tokenId
      - fromAccountIndex
      - amount
      - fee
      - nonce
      - signature
...
    PoolL2Transactions:
      type: object
      properties:
        transactions:
          type: array
          description: List of pool l2 transactions
          items:
            $ref: '#/components/schemas/PoolL2Transaction'
        pendingItems:
          $ref: '#/components/schemas/PendingItems'
      required:
        - transactions
        - pendingItems
      additionalProperties: false
...
    TransactionId:
      type: string
      description: Identifier for transactions. Used for any kind of transaction (both L1 and L2). More info on how the identifiers are built [here](https://idocs.hermez.io/#/spec/architecture/db/README?id=txid)
      example: "0x00000000000001e240004700" 
...
    EthereumAddress:
      type: string
      description: "Address of an Ethereum account."
      pattern: "^0x[a-fA-F0-9]{40}$"
      example: "0xaa942cfcd25ad4d90a62358b0dd84f33b398262a" 
    HezEthereumAddress:
      type: string
      description: "Address of an Ethereum account linked to the Hermez Network."
      pattern: "^hez:0x[a-fA-F0-9]{40}$"
      example: "hez:0xaa942cfcd25ad4d90a62358b0dd84f33b398262a" 
    BJJ:
      type: string
      description: "BabyJubJub compressed public key, encoded as base64 URL (RFC 4648), which result in 33 bytes. The padding byte is replaced by a sum of the encoded bytes."
      pattern: "^hez:[A-Za-z0-9_-]{44}$"
      example: "hez:rR7LXKal-av7I56Y0dEBCVmwc9zpoLY5ERhy5w7G-xwe"
    AccountIndex:
      type: string
      description: >-
        Identifier of an account. It references the position where the account is inside the state Merkle tree.
        The identifier is built using: `hez:` + `token symbol:` + `index`
      example: "hez:DAI:4444"
    TransactionType:
      type: string
      description: Type of transaction.
      enum:
      - Exit
      - Transfer
      - Deposit
      - CreateAccountDeposit
      - CreateAccountDepositTransfer
      - DepositTransfer
      - ForceTransfer
      - ForceExit
      - TransferToEthAddr
      - TransferToBJJ
    TransactionTypeL2:
      type: string
      description: Type of transaction.
      enum:
      - Exit
      - Transfer
      - TransferToEthAddr
      - TransferToBJJ
    TokenId:
      type: integer
      description: Identifier of a token registered in the network.
      minimum: 0
      maximum: 4294967295
      example: 98765
    BigInt:
      type: string
      description: BigInt is an integer encoded as a string for numbers that are very large.
      example: "8708856933496328593"
      pattern: "^\\d+$"
    FeeSelector:
      type: integer
      description: Index of the fee type to select, more info [here](https://idocs.hermez.io/#/spec/zkrollup/fee-table?id=transaction-fee-table).
      minimum: 0
      maximum: 256
      example: 36
    Nonce:
      type: integer
      description: Number that can only be used once per account, increments by one at each transaction.
      minimum: 0
      maximum: 1.84467440737096e+19
      example: 121
    PoolL2TransactionState:
      type: string
      description: > 
        State of a L2 transaction from the coordinator pool.
          * pend: Pending
          * fing: Forging
          * fged: Forged
          * invl: Invalid  
      enum:
      - pend
      - fing
      - fged
      - invl    
...
    BJJSignature:
      type: string
      description: BabyJubJub compressed signature.
      pattern: "^[a-fA-F0-9]{128}$"
      example: "72024a43f546b0e1d9d5d7c4c30c259102a9726363adcc4ec7b6aea686bcb5116f485c5542d27c4092ae0ceaf38e3bb44417639bd2070a58ba1aa1aab9d92c03"
...
    Token:
      type: object
      description: Hermez Network compatible and registered token.
      properties:
        id:
          $ref: '#/components/schemas/TokenId'
        ethereumAddress:
          allOf:
            - $ref: '#/components/schemas/EthereumAddress'
            - description: Ethereum address in which the token is deployed.
            - example: "0xaa942cfcd25ad4d90a62358b0dd84f33b398262a"
        itemId:
          $ref: '#/components/schemas/ItemId'
        name:
          type: string
          description: full name of the token
          example: Maker Dai
        symbol:
          allOf:
            - $ref: '#/components/schemas/TokenSymbol'
            - example: DAI
        decimals:
          type: integer
          description: Number of decimals of the token.
          example: 18
        ethereumBlockNum:
          allOf:
            - $ref: '#/components/schemas/EthBlockNum'
            - description: Ethereum block number in which the token was added to the Hermez Network.
            - example: 539847538
        USD:
          type: number
          description: Value of the token in USD.
          example: 1.01
          nullable: true
        fiatUpdate:
          type: string
          format: date-time
          description: Timestamp of the moment the `USD` value was updated.
          nullable: true
      required:
        - id
        - ethereumAddress
        - itemId
        - name
        - symbol
        - decimals
        - ethereumBlockNum
        - USD
        - fiatUpdate
      additionalProperties: false
...
    Account:
      type: object
      description: State tree leaf. It contains balance and nonce of an account.
      properties:
        itemId:
          $ref: '#/components/schemas/ItemId'
        accountIndex:
          $ref: '#/components/schemas/AccountIndex'
        nonce:
          $ref: '#/components/schemas/Nonce'      
        balance:
          $ref: '#/components/schemas/BigInt'
        bjj:
          $ref: '#/components/schemas/BJJ'      
        hezEthereumAddress:
          $ref: '#/components/schemas/HezEthereumAddress'
        token:
          $ref: '#/components/schemas/Token'
      example:
        accountIndex: hez:ETH:259
        balance: '2590000000000000000'
        bjj: hez:W6x4TZOAZ9mAqdOb3Xm_hKDLspaXfEfMMN4tXOkinS-W
        hezEthereumAddress: hez:0x7E5F4552091A69125d5DfCb7b8C2659029395Bdf
        itemId: 4
        nonce: 0
        token:
          USD: 500
          decimals: 18
          ethereumAddress: '0x0000000000000000000000000000000000000000'
          ethereumBlockNum: 0
          fiatUpdate: '2020-11-26T09:53:47.444444Z'
          id: 0
          itemId: 1
          name: Ether
          symbol: ETH
      additionalProperties: false
      required:
        - itemId
        - accountIndex
        - nonce
        - balance
        - bjj
        - hezEthereumAddress
        - token
    Accounts:
      type: object
      properties:
        accounts:
          type: array
          description: List of accounts.
          items: 
            $ref: '#/components/schemas/Account'
        pendingItems:
          $ref: '#/components/schemas/PendingItems'
      example:
        accounts:
        - accountIndex: hez:ETH:259
          balance: '2590000000000000000'
          bjj: hez:W6x4TZOAZ9mAqdOb3Xm_hKDLspaXfEfMMN4tXOkinS-W
          hezEthereumAddress: hez:0x7E5F4552091A69125d5DfCb7b8C2659029395Bdf
          itemId: 4
          nonce: 0
          token:
            USD: 500
            decimals: 18
            ethereumAddress: '0x0000000000000000000000000000000000000000'
            ethereumBlockNum: 0
            fiatUpdate: '2020-11-26T09:53:47.444444Z'
            id: 0
            itemId: 1
            name: Ether
            symbol: ETH
        - accountIndex: hez:TKN1:261
          balance: '2610000000'
          bjj: hez:W6x4TZOAZ9mAqdOb3Xm_hKDLspaXfEfMMN4tXOkinS-W
          hezEthereumAddress: hez:0x7E5F4552091A69125d5DfCb7b8C2659029395Bdf
          itemId: 6
          nonce: 0
          token:
            USD: 423
            decimals: 18
            ethereumAddress: '0x0000000000000000000000000000000000000064'
            ethereumBlockNum: 2
            fiatUpdate: '2020-11-26T09:53:47.456619Z'
            id: 1
            itemId: 2
            name: Test Token 1
            symbol: TKN1
        pendingItems: 0
      additionalProperties: false
      required:
        - accounts
        - pendingItems
...
    PendingItems:
      type: integer
      description: Amount of items that will be returned in subsequent calls to the endpoint, as long as they are done with same filters. When the value is 0 it means that all items have been sent.
      example: 15
...
    Error400:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Invalid signature.
            code: 4
            type: ErrInvalidSignature
    Error404:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Item(s) not found.
    Error500:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Database error.
...